Fix Calendar circular types
