+++ packages/components/src/Calendar/utils/setFocusInCalendar.spec.tsx
@@ -3,7 +3,8 @@ import { render, screen, within } from '@testing-library/react'
 import { format } from 'date-fns'
 import { enUS } from 'date-fns/locale'
 import { vi } from 'vitest'
-import { CalendarSingle, type CalendarSingleProps } from '../CalendarSingle'
+import { CalendarSingle } from '../CalendarSingle'
+import { type CalendarSingleProps } from '../types'
 import { setFocusInCalendar } from './setFocusInCalendar'
 const CalendarWrapper = (props: Partial<CalendarSingleProps>): JSX.Element => (
   <CalendarSingle
